Create folder & file for GenerateCards
